Fixed: Problemas menores Problemas detectados los test unitarios
Añadido:
Test unitarios

Modificado:
Las categorías asociadas a las nota ahora coinciden con el id y no con el nombre de la categoria

+++ app/src/main/java/com/android/demo/notepad3/CategoryList.java
@@ -112,7 +112,9 @@ public class CategoryList extends ListActivity {
         switch(item.getItemId()) {
             case DELETE_CATEGORY:
                 AdapterContextMenuInfo info = (AdapterContextMenuInfo) item.getMenuInfo();
-                mDbHelper.deleteCategory(info.id);
+                if(info.id != CategoriesDbAdapter.getDefault_category()) {
+                    mDbHelper.deleteCategory(info.id);
+                }
                 fillData();
                 return true;
             case SHOW_NOTES_THIS:
